Moved common methods related to the request to its own file.
