+++ src/lib/profileLoader.ts
@@ -1,4 +1,4 @@
-import { Base, Helper } from "../utils";
+import { Base, Request } from "../utils";
 import { IProfileLoader } from "./types";
 import { ITargetInfoProps } from "../utils/types";
 
@@ -14,7 +14,7 @@ export const ProfileLoader: IProfileLoader = ((targetInfo?: ITargetInfoProps) =>
     profileLoader.targetInfo.method = "POST";
 
     // Add the methods
-    Helper.addMethods(profileLoader, { __metadata: { type: "profileloader" } });
+    Request.addMethods(profileLoader, { __metadata: { type: "profileloader" } });
 
     // Return the profile loader
     return profileLoader;
