+++ src/lib/utility.ts
@@ -1,5 +1,5 @@
 import * as Types from "../mapper/types";
-import { Base, Helper, RequestType } from "../utils";
+import { Base, Request, RequestType } from "../utils";
 import { IUtility } from "./types";
 
 /**
@@ -19,7 +19,7 @@ export const Utility: IUtility = ((url?, targetInfo?) => {
     }
 
     // Add the methods
-    Helper.addMethods(utility as any, { __metadata: { type: "utility" } });
+    Request.addMethods(utility as any, { __metadata: { type: "utility" } });
 
     // Method to create a wiki page
     utility.createWikiPage = (listUrl: string, content: string = "") => {
