+++ src/utils/types/index.d.ts
@@ -1,6 +1,7 @@
 export * from "./base";
 export * from "./helper";
 export * from "./methodInfo";
+export * from "./request";
 export * from "./requestType";
 export * from "./targetInfo";
 export * from "./xhrRequest";
